refactor: move resolvers up
