+++ packages/@tissuumaps-render/src/resolvers/ColorResolver.ts
