+++ packages/@tissuumaps-render/src/resolvers/OpacityResolver.ts
